remove combined view and upgrade normal graph to better fir forecast
