+++ Trio Watch App Extension/Helper/Helper+Enums.swift
@@ -67,81 +67,3 @@ enum WatchSize {
         }
     }
 }
-
-// MARK: - Per-device UI metrics
-
-extension WatchSize {
-    /// Diameter of the glucose circle ring.
-    var circleSize: CGFloat {
-        switch self {
-        case .watch40mm: return 82
-        case .watch41mm,
-             .watch42mm: return 86
-        case .watch44mm: return 96
-        case .unknown,
-             .watch45mm: return 103
-        case .watch49mm: return 105
-        }
-    }
-
-    /// Stroke width of the glucose circle ring.
-    var lineWidth: CGFloat {
-        switch self {
-        case .watch40mm,
-             .watch41mm,
-             .watch42mm,
-             .watch44mm: return 1
-        case .unknown,
-             .watch45mm,
-             .watch49mm: return 1.5
-        }
-    }
-
-    /// Glow / shadow radius of the glucose circle ring.
-    var shadowRadius: CGFloat {
-        switch self {
-        case .watch40mm,
-             .watch41mm,
-             .watch42mm: return 8
-        case .watch44mm: return 9
-        case .unknown,
-             .watch45mm,
-             .watch49mm: return 12
-        }
-    }
-
-    /// Font size for the current glucose reading inside the circle.
-    var currentGlucoseFontSize: Font {
-        switch self {
-        case .watch40mm,
-             .watch41mm,
-             .watch42mm,
-             .watch44mm: return .title2
-        case .unknown,
-             .watch45mm,
-             .watch49mm: return .title
-        }
-    }
-
-    /// Font size for the "X m ago" label below the circle (GlucoseTrendView).
-    var minutesAgoFontSize: CGFloat {
-        switch self {
-        case .watch40mm,
-             .watch41mm: return 9
-        case .unknown,
-             .watch42mm,
-             .watch44mm: return 10
-        case .watch45mm: return 11
-        case .watch49mm: return 10
-        }
-    }
-
-    /// Scale factor used when rendering the minimized circle in CombinedGlucoseChartview.
-    var minimizedScale: CGFloat { 0.50 }
-
-    /// Width of the spacer that flanks the minimized circle in CombinedGlucoseChartview,
-    /// keeping the adjacent texts a consistent 4 pt away from the circle edge on every watch.
-    var minimizedCircleSpacerWidth: CGFloat {
-        circleSize * minimizedScale + 8 // 4 pt gap on each side
-    }
-}
