+++ Trio Watch App Extension/Views/GlucoseChartView.swift
@@ -7,8 +7,6 @@ import SwiftUI
 struct GlucoseChartView: View {
     let state: WatchState
     let glucoseValues: [(date: Date, glucose: Double, color: Color)]
-    let minYAxisValue: Decimal
-    let maxYAxisValue: Decimal
 
     @State private var timeWindow: TimeWindow = .threeHours
 
@@ -28,19 +26,38 @@ struct GlucoseChartView: View {
         }
     }
 
-    // TODO: should we only change the x axis here like we do in the main chart instead of filtering the values?
-    private var filteredValues: [(date: Date, glucose: Double, color: Color)] {
-        let cutoffDate = Date().addingTimeInterval(-Double(timeWindow.rawValue) * 3600)
-        return glucoseValues.filter { $0.date > cutoffDate }
+    private var xAxisDomain: ClosedRange<Date> {
+        let now = Date()
+        let pastOffset = -Double(timeWindow.rawValue - 2) * 3600
+        let startDate = now.addingTimeInterval(pastOffset)
+
+        // By default, the chart ends at "now" if no forecast is active
+        var endDate = now
+
+        if state.showForecast {
+            if state.isForecastCone {
+                if let lastConePoint = forecastConePoints.last {
+                    endDate = lastConePoint.date
+                } else {
+                    endDate = now.addingTimeInterval(2 * 3600) // Fallback if data is missing
+                }
+            } else {
+                if let maxLineDate = forecastLinePoints.map(\.date).max() {
+                    endDate = maxLineDate
+                } else {
+                    endDate = now.addingTimeInterval(2 * 3600) // Fallback if data is missing
+                }
+            }
+        }
+
+        // Prevents an invalid range interval if the calculated end date is in the past
+        return startDate ... max(now, endDate)
     }
 
-    var glucosePointSize: CGFloat {
-        switch timeWindow {
-        case .threeHours: return 18
-        case .sixHours: return 14
-        case .twelveHours: return 10
-        case .twentyFourHours: return 6
-        }
+    private var filteredValues: [(date: Date, glucose: Double, color: Color)] {
+        let pastHours = Double(timeWindow.rawValue - 2)
+        let cutoffDate = Date().addingTimeInterval(-pastHours * 3600)
+        return glucoseValues.filter { $0.date > cutoffDate }
     }
 
     struct ForecastConePoint: Identifiable {
@@ -92,8 +109,56 @@ struct GlucoseChartView: View {
         return pts
     }
 
+    var glucosePointSize: CGFloat {
+        switch timeWindow {
+        case .threeHours: return 18
+        case .sixHours: return 14
+        case .twelveHours: return 10
+        case .twentyFourHours: return 6
+        }
+    }
+
+    private var yAxisBounds: (min: Double, max: Double)? {
+        let values = filteredValues.map(\.glucose) +
+            forecastConePoints.flatMap { [$0.min, $0.max] } +
+            forecastLinePoints.map(\.value)
+
+        guard let minValue = values.min(),
+              let maxValue = values.max()
+        else {
+            return nil
+        }
+        return (minValue, maxValue)
+    }
+
+    private var yAxisDomain: ClosedRange<Double> {
+        guard let bounds = yAxisBounds else { return 0 ... 1 }
+        let padding = max((bounds.max - bounds.min) * 0.20, 10)
+        return (bounds.min - padding) ... (bounds.max + padding)
+    }
+
+    private var yAxisValues: [Double] {
+        guard let bounds = yAxisBounds else { return [] }
+        guard bounds.min != bounds.max else { return [bounds.min] }
+        let middle = roundedUpMiddle(for: bounds)
+        return [bounds.min, middle, bounds.max].reduce(into: [Double]()) { values, value in
+            guard !values.contains(where: { abs($0 - value) < 0.0001 }) else { return }
+            values.append(value)
+        }
+    }
+
+    private func roundedUpMiddle(for bounds: (min: Double, max: Double)) -> Double {
+        let step = bounds.max < 40 ? 0.1 : 1
+        let middle = (bounds.min + bounds.max) / 2
+        return ceil(middle / step) * step
+    }
+
+    private func formattedYAxisLabel(for glucose: Double) -> String {
+        glucose < 40 ? String(format: "%.1f", glucose) : "\(Int(glucose))"
+    }
+
     var body: some View {
-        VStack(spacing: 8) {
+        VStack(spacing: 0) {
             if filteredValues.isEmpty {
                 Text("No glucose readings.").font(.headline)
                 Text("Check phone and CGM connectivity.").font(.caption)
@@ -110,6 +175,7 @@ struct GlucoseChartView: View {
 
                     if state.showForecast {
                         if state.isForecastCone {
+                            // Only the fill area of the cone — no border lines for min/max.
                             ForEach(forecastConePoints) { pt in
                                 AreaMark(
                                     x: .value("Time", pt.date),
@@ -119,6 +185,7 @@ struct GlucoseChartView: View {
                                 .foregroundStyle(Color.insulin.opacity(0.3))
                             }
                         } else {
+                            // Uses the cleanly capped forecast line points in the chart
                             ForEach(forecastLinePoints) { pt in
                                 LineMark(
                                     x: .value("Time", pt.date),
@@ -139,22 +206,23 @@ struct GlucoseChartView: View {
                 ])
                 .chartLegend(.hidden)
                 .chartXAxis(.hidden)
+                .chartXScale(domain: xAxisDomain)
                 .chartYAxisLabel("\(timeWindow.rawValue) h", alignment: .topLeading)
                 .chartYAxis {
-                    AxisMarks(position: .trailing) { value in
+                    AxisMarks(position: .trailing, values: yAxisValues) { value in
                         AxisGridLine(stroke: .init(lineWidth: 0.65, dash: [2, 3]))
                             .foregroundStyle(Color.white.opacity(0.25))
 
                         AxisValueLabel {
                             if let glucose = value.as(Double.self) {
-                                Text("\(Int(glucose))")
+                                Text(formattedYAxisLabel(for: glucose))
+                                    .font(.caption2)
+                                    .foregroundStyle(.secondary)
                             }
                         }
                     }
                 }
-                .chartYScale(
-                    domain: minYAxisValue ... maxYAxisValue
-                )
+                .chartYScale(domain: yAxisDomain)
                 .chartPlotStyle { plotContent in
                     plotContent
                         .background(
@@ -163,10 +231,12 @@ struct GlucoseChartView: View {
                         )
                         .clipShape(RoundedRectangle(cornerRadius: 12))
                 }
-                .padding(.bottom)
+                // Allow the chart to expand fully to utilize all vertical room
+                .frame(maxHeight: .infinity)
             }
         }
-        .scenePadding()
+        // Switched from scenePadding() to horizontal-only padding to let the chart stretch taller
+        .padding(.horizontal, 6)
         .onTapGesture {
             withAnimation {
                 timeWindow = timeWindow.next
