+++ Trio Watch App Extension/Views/TrioMainWatchView.swift
@@ -62,14 +62,7 @@ struct TrioMainWatchView: View {
     var body: some View {
         NavigationStack(path: $navigationPath) {
             TabView(selection: $currentPage) {
-                // Page 1: Combined View of glucose trend and chart view
-                CombinedGlucoseChartview(
-                    state: state,
-                    rotationDegrees: rotationDegrees,
-                    isWatchStateDated: isWatchStateDated || isSessionUnreachable
-                )
-                .tag(0)
-                // Page 2: Current glucose trend in "BG bobble"
+                // Page 1: Current glucose trend in "BG bobble"
                 ZStack {
                     GlucoseTrendView(
                         state: state,
@@ -92,16 +85,14 @@ struct TrioMainWatchView: View {
                                     7 // Font .body == 14, so half of default size for the SF Symbol image
                             )
                     }
-                }.tag(1)
+                }.tag(0)
 
-                // Page 3: Glucose chart
+                // Page 2: Glucose chart
                 GlucoseChartView(
                     state: state,
-                    glucoseValues: state.glucoseValues,
-                    minYAxisValue: state.minYAxisValue,
-                    maxYAxisValue: state.maxYAxisValue
+                    glucoseValues: state.glucoseValues
                 )
-                .tag(2)
+                .tag(1)
             }
             .onAppear {
                 /// Hard reset variables when main view appears
